requirements.txt
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
backcall==0.2.0
browser-cookie3==0.20.1
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
cryptography==41.0.7
colorama==0.4.6
decorator==5.2.1
Deprecated==1.2.18
dj-database-url==2.3.0
Django==5.2.4
django-allauth==65.11.0
django-debug-toolbar==4.4.6
django-extensions==3.1.5
djangorestframework==3.14.0
frozenlist==1.7.0
future==1.0.0
google-auth==2.40.3
google-genai==1.26.0
gunicorn==21.2.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
ipython==7.34.0
jedi==0.19.2
line-bot-sdk==3.17.1
loguru==0.7.3
lz4==4.4.4
matplotlib-inline==0.1.7
multidict==6.6.3
packaging==25.0
parso==0.8.4
pickleshare==0.7.5
Pillow==10.0.1
prompt_toolkit==3.0.51
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.8.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
rsa==4.9.1
shadowcopy==0.0.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tenacity==8.5.0
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.1
